requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
openpyxl>=3.1.2
pandas>=2.2.0
python-dateutil>=2.8.2
anthropic>=0.39.0
pdf2image>=1.17.0
Pillow>=10.1.0
PyPDF2>=3.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
